+++ lib/src/model/email.dart
@@ -7,8 +7,8 @@ class Email {
   Email({this.id = '', this.domain = ''});
 
   @override
+  // ignore: type_annotate_public_apis
   bool operator ==(other) {
-    // ignore: type_annotate_public_apis
     if (other is Email) {
       return other.id == id && other.domain == domain;
     }