+++ lib/src/model/gpx.dart
@@ -14,10 +14,12 @@ class Gpx {
   List<Wpt> wpts = [];
   List<Rte> rtes = [];
   List<Trk> trks = [];
+  
+  Map<String, String> extensions = {};
 
   @override
+  // ignore: type_annotate_public_apis
   bool operator ==(other) {
-    // ignore: type_annotate_public_apis
     if (other is Gpx) {
       return other.creator == creator &&
           other.version == version &&