+++ lib/src/model/kml_tag.dart
@@ -1,0 +1,30 @@
+class KmlTagV22 {
+  static const kml = 'kml';
+
+  static const document = 'Document';
+
+  static const placemark = 'Placemark';
+  static const name = 'name';
+  static const desc = 'description';
+
+  static const point = 'Point';
+  static const track = 'LineString';
+  static const coordinates = 'coordinates';
+
+  static const extendedData = 'ExtendedData';
+  static const data = 'Data';
+
+  static const value = 'value';
+
+  static const altitudeMode = 'altitudeMode';
+  static const extrude = 'extrude';
+  static const tessellate = 'tessellate';
+
+  static const timestamp = 'timestamp';
+  static const when = 'when';
+
+  // <altitudeMode>absolute</altitudeMode>
+
+
+
+}