+++ lib/src/model/rte.dart
@@ -29,8 +29,8 @@ class Rte {
         rtepts = rtepts ?? [];
 
   @override
+  // ignore: type_annotate_public_apis
   bool operator ==(other) {
-    // ignore: type_annotate_public_apis
     if (other is Rte) {
       return other.name == name &&
           other.cmt == cmt &&