+++ lib/src/model/trk.dart
@@ -28,8 +28,8 @@ class Trk {
         trksegs = trksegs ?? [];
 
   @override
+  // ignore: type_annotate_public_apis
   bool operator ==(other) {
-    // ignore: type_annotate_public_apis
     if (other is Trk) {
       return other.name == name &&
           other.cmt == cmt &&