+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+## [0.0.3] 
+
+* fix lint issues
+* add kml writer (export from Gpx into KML)
+
 ## [0.0.2] 
 
 * fix lint issues
 
 ## [0.0.1] 
 
-* Initial release - reader, writer for gpx
+* Initial release - reader, writer for GPX-files